Rotates entities on widgets into the light
This rotates entity icons on widgets (tasks, and in theory, rewards) -45 instead of 45.

This faces the entity into the light, instead of into shadow.

For villager professions, the 'face' face of their textures, generall, has key visually distingishing features.
Hence, this change improves their legibility, especially they are rendered small enough to fit the widget.

An alternate fix, probably better but unknown to me how or if is possible, would be to move the light 90 degrees.
Would love to learn how if so.