Better Centers Hermes' DocumentWidget
Tested at UI scales 2, 3, and 4.

When there are no tasks or rewards (and hence, no left margin), centering is dead on.
When there is a left margin...
	...and UI scale 3: centering is spot on.
	...and UI scale is 2 or 4: centering is off by one font "pixel"; a forgivable amount for now.

Undocumented magic numbers are bad, though.
How Hermes' is centering text, <img> tags, and horizontal rules is a factor.
(the later were measured down to the pixel in screen shots as a stand-in for Hermes' position and width)